requirements.txt
Django==4.2.3
environs[django]==9.5.0
django-model-utils==4.3.1
django-allauth==0.54.0
crispy-bootstrap5==0.7
factory-boy==3.2.1
flake8==6.0.0
isort==5.12.0
channels[daphne]==4.0.0
redis==4.6.0
channels-redis==4.1.0
whitenoise==6.5.0